# netdisk-docs/openapi.yaml
openapi: 3.0.3
info:
  title: NetDisk API
  version: 1.0.0
  description: |
    类百度网盘 - MVP 可落地接口定义（Spring Boot + Vue3）。
    - 认证：Bearer JWT
    - 上传：S3/MinIO Multipart + Presigned URL
servers:
  - url: http://localhost:8080
    description: local
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录（可选：token 拉黑）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/files/folders:
    post:
      tags: [Files]
      summary: 新建文件夹
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_CreateFolderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/files/list:
    get:
      tags: [Files]
      summary: 列表目录
      parameters:
        - in: query
          name: parent_id
          schema: { type: integer, format: int64, nullable: true }
          description: 父目录ID；根目录传空或 0（实现任选其一，需前后端统一）
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: sort
          schema: { type: string, enum: [name, updated_at, size_bytes], default: updated_at }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: include_deleted
          schema: { type: integer, enum: [0, 1], default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PageEntryVO"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/files/{entryId}/rename:
    post:
      tags: [Files]
      summary: 重命名
      parameters:
        - in: path
          name: entryId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/files/move:
    post:
      tags: [Files]
      summary: 移动（批量）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/files/delete:
    post:
      tags: [Files]
      summary: 删除（进回收站）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/trash/list:
    get:
      tags: [Trash]
      summary: 回收站列表
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PageEntryVO"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/trash/restore:
    post:
      tags: [Trash]
      summary: 还原
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Unauthorized:
      description: 未登录或 token 失效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    Conflict:
      description: 冲突（同名/状态冲突/幂等冲突等）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    ApiResponse_Object:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: OK }
        data: { type: object, nullable: true }
        request_id: { type: string, example: req_xxx }
    ApiError:
      type: object
      properties:
        code: { type: integer, example: 40001 }
        message: { type: string, example: 参数错误 }
        request_id: { type: string, example: req_xxx }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: a@b.com }
        password: { type: string, example: "12345678" }
    RegisterResponse:
      type: object
      properties:
        user_id: { type: integer, format: int64 }
    ApiResponse_RegisterResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: OK }
        data: { $ref: "#/components/schemas/RegisterResponse" }
        request_id: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
        expire_in: { type: integer, example: 7200 }
    ApiResponse_LoginResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: OK }
        data: { $ref: "#/components/schemas/LoginResponse" }
        request_id: { type: string }
    CreateFolderRequest:
      type: object
      required: [parent_id, name]
      properties:
        parent_id: { type: integer, format: int64, example: 0 }
        name: { type: string, example: 新建文件夹 }
    CreateFolderResponse:
      type: object
      properties:
        entry_id: { type: integer, format: int64 }
    ApiResponse_CreateFolderResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: OK }
        data: { $ref: "#/components/schemas/CreateFolderResponse" }
        request_id: { type: string }
    EntryVO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        parent_id: { type: integer, format: int64, nullable: true }
        type: { type: string, enum: [folder, file] }
        name: { type: string }
        size_bytes: { type: integer, format: int64 }
        is_deleted: { type: boolean }
        deleted_at: { type: string, nullable: true }
        created_at: { type: string }
        updated_at: { type: string }
    PageEntryVO:
      type: object
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/EntryVO" }
    ApiResponse_PageEntryVO:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: OK }
        data: { $ref: "#/components/schemas/PageEntryVO" }
        request_id: { type: string }
    RenameRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
    MoveRequest:
      type: object
      required: [entry_ids, target_parent_id]
      properties:
        entry_ids:
          type: array
          items: { type: integer, format: int64 }
        target_parent_id: { type: integer, format: int64 }
    DeleteRequest:
      type: object
      required: [entry_ids]
      properties:
        entry_ids:
          type: array
          items: { type: integer, format: int64 }
    RestoreRequest:
      type: object
      required: [entry_ids, strategy]
      properties:
        entry_ids:
          type: array
          items: { type: integer, format: int64 }
        strategy:
          type: string
          enum: [auto_rename, overwrite, fail]
...